+++ src/main/java/com/epam/reportportal/cucumber/AbstractReporter.java
@@ -162,10 +162,9 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	}
 
 	protected void startRule(@Nonnull RunningContext.FeatureContext featureContext, @Nonnull RunningContext.RuleContext ruleContext) {
-		Maybe<String> id = Utils.startNonLeafNode(
-				launch.get(),
+		Maybe<String> id = Utils.startNonLeafNode(launch.get(),
 				featureContext.getFeatureId(),
-				Utils.buildNodeName(ruleContext.getKeyword(), AbstractReporter.COLON_INFIX, ruleContext.getName(), null),
+				Utils.buildName(ruleContext.getKeyword(), AbstractReporter.COLON_INFIX, ruleContext.getName()),
 				ruleContext.getDescription(),
 				getCodeRef(ruleContext.getUri(), ruleContext.getLine()),
 				ruleContext.getAttributes(),
@@ -216,8 +215,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		}
 
 		Launch myLaunch = launch.get();
-		Maybe<String> id = Utils.startNonLeafNode(
-				myLaunch,
+		Maybe<String> id = Utils.startNonLeafNode(myLaunch,
 				rule == null ? featureContext.getFeatureId() : rule.getId(),
 				scenarioName,
 				description,
@@ -298,7 +296,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	 */
 	protected StartTestItemRQ buildStartStepRequest(TestStep testStep, String stepPrefix, String keyword) {
 		StartTestItemRQ rq = new StartTestItemRQ();
-		rq.setName(Utils.buildNodeName(stepPrefix, keyword, Utils.getStepName(testStep), ""));
+		rq.setName(Utils.buildName(stepPrefix, keyword, Utils.getStepName(testStep)));
 		rq.setDescription(Utils.buildMultilineArgument(testStep));
 		rq.setStartTime(Calendar.getInstance().getTime());
 		rq.setType("STEP");
@@ -490,7 +488,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		StartTestItemRQ rq = new StartTestItemRQ();
 		rq.setDescription(getDescription(context.getUri()));
 		rq.setCodeRef(getCodeRef(context.getUri(), 0));
-		rq.setName(Utils.buildNodeName(featureKeyword, AbstractReporter.COLON_INFIX, featureName, null));
+		rq.setName(Utils.buildName(featureKeyword, AbstractReporter.COLON_INFIX, featureName));
 		rq.setAttributes(context.getAttributes());
 		rq.setStartTime(Calendar.getInstance().getTime());
 		rq.setType(getFeatureTestItemType());
@@ -581,12 +579,7 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 		}
 
 		RunningContext.ScenarioContext newScenarioContext = featureContext.getScenarioContext(testCase);
-		String scenarioName = Utils.buildNodeName(
-				newScenarioContext.getKeyword(),
-				AbstractReporter.COLON_INFIX,
-				newScenarioContext.getName(),
-				newScenarioContext.getOutlineIteration()
-		);
+		String scenarioName = Utils.buildName(newScenarioContext.getKeyword(), AbstractReporter.COLON_INFIX, newScenarioContext.getName());
 
 		Pair<Integer, URI> scenarioLineFeatureURI = Pair.of(newScenarioContext.getLine(), featureContext.getUri());
 		RunningContext.ScenarioContext scenarioContext = currentScenarioContextMap.computeIfAbsent(scenarioLineFeatureURI, k -> {
@@ -620,16 +613,14 @@ public abstract class AbstractReporter implements ConcurrentEventListener {
 	}
 
 	protected void addToTree(RunningContext.ScenarioContext scenarioContext, String text, Maybe<String> stepId) {
-		retrieveLeaf(
-				scenarioContext.getFeatureUri(),
+		retrieveLeaf(scenarioContext.getFeatureUri(),
 				scenarioContext.getLine(),
 				ITEM_TREE
 		).ifPresent(scenarioLeaf -> scenarioLeaf.getChildItems().put(createKey(text), TestItemTree.createTestItemLeaf(stepId, 0)));
 	}
 
 	protected void removeFromTree(RunningContext.ScenarioContext scenarioContext, String text) {
-		retrieveLeaf(
-				scenarioContext.getFeatureUri(),
+		retrieveLeaf(scenarioContext.getFeatureUri(),
 				scenarioContext.getLine(),
 				ITEM_TREE
 		).ifPresent(scenarioLeaf -> scenarioLeaf.getChildItems().remove(createKey(text)));